🐛 Override Bulkrax to handle nil values from hash
This commit will handle nil values from the hash caused by the csv
having a header but no value.  We might want to put this upstream but I
seem to remember hearing of some decision why `nil` values were being
kept.  Not sure if that is still current though but at least it'll be
fixed here.

Ref:
- https://github.com/notch8/wvu_knapsack/issues/74