Revert overrides for handling blank fields
The M3 profile had specified data_type: string for certain fields, and
this caused issues when those fields were left blank. It is assumed that
the default data type is array, so when these weren't an array they were
being converted to arrays with a single blank entry. By removing these
overrides, blank fields will now be handled correctly.

We may want to revisit how strings are used in Hyrax/Hyku to prevent
similar issues in the future.